Refactor CLI into command-per-file structure
